refactor(team-mode): split messaging live delivery
